Overhauled functions build_gateway() and cybernetics_core()
build_gateway() is renamed unit_production_buildings(). It's now responsible for all offensive unit production buildings.

cybernetics_core() function will be redefined to a research function later on, and is left as a function stub.

This also fixes the bugs where it would sometimes create 2 gateways or 2 cybernetic cores. Through the black magic that is "if self.iteration % 6 == 0:"